requirements.txt
click
python-dotenv
google-generativeai
Pillow
PyYAML
tenacity
# fastapi
# uvicorn
pytest